refactor(di): extract shared parse_cookies into cookie_util module
Deduplicate the identical parse_cookies implementation from cookie.rs
and cookie_named.rs into a shared cookie_util module.

Closes #471
