+++ crates/reinhardt-di/src/params.rs
@@ -38,6 +38,7 @@
 
 pub mod body;
 pub mod cookie;
+pub(crate) mod cookie_util;
 pub mod cookie_named;
 pub mod extract;
 pub mod form;
